+++ com.unity.ml-agents/Editor/BehaviorParametersEditor.cs
@@ -1,10 +1,10 @@
-using MLAgents.Sensors;
+using Unity.MLAgents.Sensors;
 using UnityEditor;
 using Barracuda;
-using MLAgents.Policies;
+using Unity.MLAgents.Policies;
 using UnityEngine;
 
-namespace MLAgents.Editor
+namespace Unity.MLAgents.Editor
 {
     /*
      This code is meant to modify the behavior of the inspector on Agent Components.