+++ com.unity.ml-agents/Runtime/Policies/BehaviorParameters.cs
@@ -3,7 +3,7 @@ using System;
 using UnityEngine;
 using UnityEngine.Serialization;
 
-namespace MLAgents.Policies
+namespace Unity.MLAgents.Policies
 {
     /// <summary>
     /// Defines what type of behavior the Agent will be using