+++ com.unity.ml-agents/Runtime/Policies/BrainParameters.cs
@@ -2,7 +2,7 @@ using System;
 using UnityEngine;
 using UnityEngine.Serialization;
 
-namespace MLAgents.Policies
+namespace Unity.MLAgents.Policies
 {
     /// <summary>
     /// Whether the action space is discrete or continuous.