+++ com.unity.ml-agents/Tests/Editor/DiscreteActionOutputApplierTest.cs
@@ -2,10 +2,10 @@ using System;
 using Barracuda;
 using NUnit.Framework;
 using UnityEngine;
-using MLAgents.Inference;
-using MLAgents.Inference.Utils;
+using Unity.MLAgents.Inference;
+using Unity.MLAgents.Inference.Utils;
 
-namespace MLAgents.Tests
+namespace Unity.MLAgents.Tests
 {
     public class DiscreteActionOutputApplierTest
     {