+++ com.unity.ml-agents/CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to
 ## [1.0.0-preview] - 2020-05-06
 
 ### Major Changes
+- The `MLAgents` C# namespace was renamed to `Unity.MLAgents`, and other nested
+  namespaces were similarly renamed. (#3843)
 - Added new 3-joint Worm ragdoll environment. (#3798)
 - The `--load` and `--train` command-line flags have been deprecated. Training
   now happens by default, and use `--resume` to resume training instead. (#3705)